ft(AppLayout): add an info component globally for login redirection
